requirements.txt
fastapi==0.104
uvicorn==0.22.0
pydantic==1.10.5
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain==0.0.103
langchain_google_genai==0.0.1
faiss-cpu==1.7.2
